+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -7,6 +7,8 @@ description: GeckoView API Changelog.
 <h1> GeckoView API Changelog. </h1>
 
 ## v66
+- Added nested `ContentBlocking` runtime settings.
+
 - Added `RuntimeSettings` base class to support nested settings.
 
 - Added `baseUri` to [`ContentDelegate.ContextElement`][65.21] and changed
@@ -120,4 +122,4 @@ description: GeckoView API Changelog.
 [65.24]: ../CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: ../GeckoResult.html
 
-[api-version]: 53ffb6141c90d561251dd831e9af1bd7197c3c75
+[api-version]: 2b12249de4649178832780cb5c0aba13079803c8